It works!
Now correctly parses the player and goalie stats tables, fortunately handling any
weird data entered for player number (* or XX). It's likely still fragile, in that
it's doing a straight type conversion of things that may not be nums or integers, but
I think they seed the data with zeroes where appropriate, except for the player num,
so this may be safe.